008 parsing

+++ 008_philosophers/src/t_philo.h
@@ -6,7 +6,7 @@
 /*   By: halvarez <halvarez@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/19 16:31:20 by halvarez          #+#    #+#             */
-/*   Updated: 2022/09/19 17:52:20 by halvarez         ###   ########.fr       */
+/*   Updated: 2022/09/20 10:11:09 by halvarez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,20 +15,21 @@
 
 # include <sys/time.h>
 # include <pthread.h>
+# include "ft_shared.h"
 
 /* Structure of times ======================================================= */
 typedef struct s_times
 {
-	unsigned int	eat_time;
-	unsigned int	think_time;
-	unsigned int	sleep_time;
+	unsigned long	eat_time;
+	unsigned long	think_time;
+	unsigned long	sleep_time;
 }	t_times;
 
 /* Structure of times ======================================================= */
 typedef struct s_philo
 {
 	unsigned int	nbr;
-	unsigned int	n_of_meals;
+	unsigned long	n_of_meals;
 	t_times			times;
 }	t_philo;
 #endif
